requirements.txt
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
transformers[torch]==4.37.2
datasets==2.17.0
scipy==1.12.0
einops==0.7.0
accelerate==0.27.2
bitsandbytes==0.42.0
wikipedia==1.4.0
google-search-results==2.4.2
ibm-generative-ai==2.1.1
openai==1.12.0
python-dotenv==1.0.1
faiss-cpu==1.7.4
ipykernel==6.29.2
rouge==1.0.1
peft==0.8.2
dateparser==1.2.0
sentence-transformers==2.3.1
chromadb==0.4.22
tensorflow==2.15.0
